include avatar

+++ src/main/java/com/example/coordin8/repository/AvatarRepository.java
@@ -3,5 +3,5 @@ package com.example.coordin8.repository;
 import com.example.coordin8.entity.AvatarEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AvatarRepository extends JpaRepository<AvatarEntity, Integer> {
+public interface AvatarRepository extends JpaRepository<AvatarEntity, Long> {
 }
